organize files, support autoconf
